// src/modules/inventory/reports/yarn-store/yarn-stock-after-allocation-report/yarn-stock-after-allocation-report-type.ts
export type LotWiseYarnStockReportType = {
  COMPANY_ID: number;
  COMPANY_NAME: string;
  COMPANY_ADDRESS: string;
  LC_ID: number;
  LC_NUMBER: string;
  WORK_ORDER_ID: number;
  WORK_ORDER_NUMBER: string;
  REFERENCE_NO: string;
  BBLC_DATE: string | null;
  YARN_ID: number;
  YARN: string;
  YARN_COMPOSITION_ID: number;
  YARN_COMPO: string;
  YARN_COUNT_ID: number;
  YARN_COUNT: string;
  YARN_TYPE_ID: number;
  YARN_TYPE: string;
  YARN_BRAND_ID: number;
  BRAND_NAME: string;
  YARN_LOT_NUMBER_ID: number;
  YARN_LOT_NUMBER: string;
  FIRST_RCV_DATE: string | null;
  LAST_RCV_DATE: string | null;
  YARN_RECEIVED_QNT: number;
  YARN_ALLOCATION_QNT: number;
  AFTER_ALLO_STOCK: number;
  YARN_ISSUE_QNT: number;
  BALANCE_QUANTITY: number;
  REPORT_DATE: string | null;
  DYED_YARN_ID: number;
  YARN_ADJUSTMENT_QTY: number;
  WO_UNIT_PRICE: number;
  TOTAL_AMOUNT: number;
};